+++ pkg/key/key.go
@@ -3,6 +3,7 @@ package key
 import (
 	"fmt"
 
+	corev1alpha1v3 "github.com/giantswarm/apiextensions/v3/pkg/apis/core/v1alpha1"
 	"github.com/giantswarm/microerror"
 	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
 	capzexp "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
@@ -109,6 +110,19 @@ func ToAzureMachinePtr(v interface{}) (*capz.AzureMachine, error) {
 	customObjectPointer, ok := v.(*capz.AzureMachine)
 	if !ok {
 		return nil, microerror.Maskf(errors.WrongTypeError, "expected '%T', got '%T'", &capz.AzureMachine{}, v)
+	}
+
+	return customObjectPointer, nil
+}
+
+func ToSparkPtr(v interface{}) (*corev1alpha1v3.Spark, error) {
+	if v == nil {
+		return nil, microerror.Maskf(errors.WrongTypeError, "expected '%T', got '%T'", &corev1alpha1v3.Spark{}, v)
+	}
+
+	customObjectPointer, ok := v.(*corev1alpha1v3.Spark)
+	if !ok {
+		return nil, microerror.Maskf(errors.WrongTypeError, "expected '%T', got '%T'", &corev1alpha1v3.Spark{}, v)
 	}
 
 	return customObjectPointer, nil